Remove unused LocalStack SQS/SNS setup scripts
Deleted legacy scripts for SQS/SNS setup in LocalStack, simplifying the infrastructure code. The required logic was consolidated into the main `setup.sh`, improving maintainability. Removed related unused Kotlin classes and tests to clean up deprecated constructs.

+++ core/simulation-domain/src/main/kotlin/br/com/simulator/credit/creditas/simulationdomain/service/LoanCalculator.kt
@@ -16,7 +16,7 @@ object LoanCalculator {
       Simulation.create(
         loanAmount = LoanAmount(loanAmount),
         months = months,
-        annualRate = annualRate.toBigDecimal(),
+        annualRate = annualRate,
       )
     return simulation.calculate()
   }
